LLM-Apps/Langchain-Examples/pyproject.toml
[project]
name = "langchain-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-docling>=0.2.0",
    "langchain-groq>=0.3.2",
    "python-dotenv>=1.1.0",
]
